refactor(Purchase): Quantity를 Purchase로 이름 변경
구매 수량을 담는 역할을 하기에 Purchase로 이름 수정
단위 금액을 Constant 클래스의 상수에서 사용하도록 수정
